Add moving a task from recents to desktop
There is a current need to launch a non-running task in desktop for
overview launch case on device reboot. This cl creates a way to launch a non-running task given a taskid if the task is found in the recents list. The
transition animation will be added once overview chip to desktop animation is done.

Bug: 332889335
Test: atest DesktopTasksControllerTest
Flag: ACONFIG com.android.window.flags.enable_desktop_windowing_mode TEAMFOOD
Change-Id: I7a4ae568b3f8878f404af786546dfee6233cda7e