Transition to hub from screen off when applicable.
This change fixes the dream->off->hub transition so that hub appears
instead of lockscreen on power on.

The original change was reverted due to a performance regression on
devices where the hub doesn't exist. This new change includes a check
that the screen turned on due to a power button press, as well as a
slight refactor to avoid doing unnecessary work when communal hub is
disabled.

Bug: 339315487, 350757214
Flag: com.android.systemui.communal_hub
Test: atest CommunalSceneStartableTest
Test: atest FromDozingTransitionInteractorTest
Test: atest CommunalInteractorTest

Change-Id: I3fe34e57a14af956c0bef8d2fa04042ad01d429e