Do not show remove button unless item is selected or reordering.
Currently, the remove button shows in edit mode if an item is selected
or we are reordering. However, the definition of reordering was defined
on simply drag start. This changelist updates the behavior to make sure
we're actually dragging an item before we are considered in the
reordering state.

Test: manually verified remove button shows under correct conditions
Flag: EXEMPT bugfix
Fixes: 346337328
Change-Id: I2f5524acea494799d0f92e5c08a49f24d5428607