Enable and configure interruptions in Flexiglass
This CL enables interruptions in Flexiglass and configures the most
common interruptions that happen when face authentication kicks in while
we were transition from one scene to Bouncer.

Enabling this flag will also to have multiple transitions to happen in
parallel, which is going to be more and more important with overlays and
once we have scenes with horizontal swipes like Communal.

See b/290184746#comment13 for details and before/after videos.

Bug: 290184746
Test: Manual, see b/290184746#comment13. Unfortunately motion tests are
 not configured for a full Flexiglass set up yet so I can't unit test
 it.
Flag: com.android.systemui.scene_container
Change-Id: Ia4fd64162d8930dcd091c6616ff219a58eeae4e3