Prevents embedded activity relaunched twice when display rotates
The activity was relaunched when the display rotated because the
TF bounds were not inside the parent Task bounds at the moment,
so the swdp was changed (inherited from the parent Task).

And then the activity relaunched again due to the embedded TF bounds
were updated by the TFOrganizer afterward.

Note that the insets may still change while the display rotates,
and making the pre-V activities bounds changed and relaunched,
with another activity relaunch followed due to embedded TF bounds
change.

Bug: 294389608
Flag: EXEMPT bugfix
Test: atest WmTests:TaskFragmentTest

Change-Id: I26934398a8d41ba9fe791e485cf6983f47863f9f