Move caption insets WCT to the shell bg thread
The WCT is applied using TaskOrganizer#applyTransaction, which is
synchronous and thus blocking the transition animation on a binder
transaction. This is unnecessary given the WCT only includes inset
source changes that won't be synchronized with the animation
regardless.
This change moves that call to the shell background thread, to avoid
adding latency to the shell transition.

Bug: 346716836
Test: verify caption's WCT slice in perfetto trace no longer waits for
two-way binder transaction.
Flag: EXEMPT bugfix

Change-Id: I1e9aa7bfbb457d325300c4f9adb78455f643bfa8